added the service

+++ src/Application/Interfaces/IPropertyRentalService.cs
@@ -6,9 +6,9 @@ namespace Application.Interfaces
     public interface IPropertyRentalService
     {
         Task AddPropertyRentalAsync(long reference, AddOrUpdatePropertyRentalModel model);
-        Task<PropertyRentModel> GetPropertyRentalAsync(long rentalId);
+        Task<PropertyRentModel> GetPropertyRentalAsync(int rentalId);
         Task RemovePropertyRentalAsync(int rentalId);
-        Task UpdatePropertyRentalAsync(long rentalId, AddOrUpdatePropertyRentalModel model);
+        Task UpdatePropertyRentalAsync(int rentalId, AddOrUpdatePropertyRentalModel model);
         Task<IEnumerable<PropertyLetTypeModel>> GetPropertyLetTypesAsync();
     }
 }
